fix: Issues with layerClass and v8 runtime (#10)
I don't know the reason for this but the v8 runtime blows up when layerClass is exposed to objective-c.  It causes the MetadataBuilder::RegisterStaticMethods function to fail.

I tested this without exposing layerClass to objective-c and everything works fine.  This is key to fixing https://github.com/nstudio/nativescript-plugins/issues/12